+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for renaming variables within the scope of a predicate clause, grammar rule, or directive
 * Add support for renaming parameter variables within the scope of an entity
+* Add "Renumber variables" refactoring support for variables ending with numbers within the scope of a predicate clause or grammar rule
 * Add "Unify with new variable" refactoring support for selected terms in predicate rules and grammar rules
 * Add "Inline variable" refactoring support for replacing variable unification goals in predicate rules and grammar rules
 * Add "Wrap file contents as an object" refactoring support for converting plain Prolog files to Logtalk objects